- Added static methods. - Changed import statements to match with python2 style.
