+++ wsgi/odyssy/basic/urls.py
@@ -1,6 +1,6 @@
 from django.conf.urls import url
 
-from . import views
+import views
 
 urlpatterns = [
     url('^$', views.index, name='index-view')
