+++ wsgi/odyssy/basic/views.py
@@ -1,8 +1,8 @@
 from django.shortcuts import render
 
-from events import views
+from events.models import Event
 
 
 def index(request):
-    events_list = views.get_latest_events(3)
+    events_list = Event.get_latest_events(3)
     return render(request, 'basic/index.html', context={'events': events_list})
